fix: some Tabs improvements (#1668)
