+++ src/utilities/elements.ts
@@ -3,6 +3,22 @@
 import { MouseEvent } from 'react';
 
 export type ItemElementType = 'a' | 'button' | 'span';
+export const TABBABLE_ELEMENTS = [
+    'input:not([disabled]):not([type=hidden])',
+    'select:not([disabled])',
+    'textarea:not([disabled])',
+    'button:not([disabled])',
+    'a[href]',
+    'area[href]',
+    'summary',
+    'iframe',
+    'object',
+    'embed',
+    'audio[controls]',
+    'video[controls]',
+    '[contenteditable]',
+    '[tabindex]:not([tabindex="-1"]):not([disabled])',
+].join(':not([hidden]):not([tabindex="-1"]),');
 
 export const getItemElementType = (link?: string, onClick?: (event: MouseEvent<HTMLButtonElement>) => void) => {
     if (link) {
@@ -16,8 +32,7 @@ export const getItemElementType = (link?: string, onClick?: (event: MouseEvent<H
 
 export const checkIfContainInteractiveElements = (node: HTMLElement | null) => {
     if (node) {
-        const focusableElements = 'button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])';
-        const focusableChildren = node.querySelectorAll(focusableElements) ?? [];
+        const focusableChildren = node.querySelectorAll(TABBABLE_ELEMENTS) ?? [];
         return focusableChildren.length > 0;
     }
     return false;
